requirements.txt
django
markdown
django-markup
django-pure-pagination
django-simple-captcha
django-bootstrap3
django-registration-redux